requirements.txt
pytest>=7.0
requests>=2.28
pydantic>=1.10
allure-pytest>=2.10.0
faker
python-dotenv